Refactored code so air pollution is in the folder now. It's 1:53am. I'm gonna go to sleep now
