+++ globes/globe_earthquake.js
@@ -35,3 +35,6 @@ const myGlobe = Globe()
 fetch('//earthquake.usgs.gov/earthquakes/feed/v1.0/summary/all_day.geojson').then(res => res.json())
 .then(obj => obj.features)
 .then(equakes => myGlobe.ringsData(equakes)) // .then(equakes => console.log(equakes[0]))
+
+
+export {myGlobe};
